+++ backend/src/main/java/com/techeer/backend/api/resume/domain/Position.java
@@ -1,0 +1,12 @@
+package com.techeer.backend.api.resume.domain;
+
+import lombok.ToString;
+
+@ToString
+public enum Position {
+    BACKEND,
+    FRONTEND,
+    ANDROID,
+    IOS,
+    DEVOPS
+}
