+++ backend/src/main/java/com/techeer/backend/api/resume/domain/ResumeTechStack.java
@@ -1,0 +1,26 @@
+package com.techeer.backend.api.resume.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ResumeTechStack {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Resume resume;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private TechStack techStack;
+
+    public ResumeTechStack(Resume resume, TechStack techStack) {
+        this.resume = resume;
+        this.techStack = techStack;
+    }
+}
