+++ backend/src/main/java/com/techeer/backend/api/resume/domain/TechStack.java
@@ -1,0 +1,28 @@
+package com.techeer.backend.api.resume.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TechStack {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @OneToMany(mappedBy = "techStack")
+    private List<ResumeTechStack> resumeTechStacks = new ArrayList<>();
+
+    public TechStack(String name) {
+        this.name = name;
+    }
+}
