--- backend/src/main/java/com/techeer/backend/domain/resume/entity/Position.java
@@ -1,12 +1,0 @@
-package com.techeer.backend.domain.resume.entity;
-
-import lombok.ToString;
-
-@ToString
-public enum Position {
-    BACKEND,
-    FRONTEND,
-    ANDROID,
-    IOS,
-    DEVOPS
-}
